Squashed commit of the following:
commit 02b07d2fad9f2b011e771682761d1a37ba22d1db
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 11:39:35 2020 -0600

    Git Formatting

commit 984cad6b3bc77f30a7a8731d2b92bfdf7d3eff77
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 11:37:17 2020 -0600

    Fixed Issue with DiffCommand

commit 1cc0db2bc8cfc841867285638355ec1b99dc6162
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 11:14:24 2020 -0600

    Fixes

commit d34be10e90973f33c378ccf0881d3ae28207857c
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 10:49:09 2020 -0600

    Naming Updates

commit d021e4e7eb9119dae9795de03dde5d793612aaca
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 09:53:35 2020 -0600

    CLI Updates

commit dafd6e2cad96526bf90a4510a7a1a9aa3df62823
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 08:30:51 2020 -0600

    Better error handling

commit f1972288917a41f5d577fa612946175e506e5bc2
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 08:13:40 2020 -0600

    Validation Logic

commit f50467573447dac637066609749383fd350cbb3d
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 07:46:58 2020 -0600

    SyncTracker

commit b4bdae5d472218983ec17a12c6cf86dc9731c8db
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jul 7 07:00:41 2020 -0600

    Query by prefix from Amazon

commit efc41a8541949eb05dc782d9e18d4771a8804afd
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Mon Jul 6 16:51:16 2020 -0600

    Now using Settings.Drive & only scanning local included directories

commit bb530aad76a6b309f08736adfe07cfed28123c31
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Mon Jul 6 11:09:09 2020 -0600

    Filename Normalization

commit 3f01f71b017b02e770e84147e418623feae5f435
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Mon Jul 6 08:02:52 2020 -0600

    Refactored to use more DI. Added updated monitor command.

commit 556a901a009eb7914c8914b865e94d6c0dc43c04
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 16:23:46 2020 -0600

    Sync Support

commit 45d912f18d998bd4e7ccc5d5ccf8a284d4c1f185
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 15:23:16 2020 -0600

    Sync Command (Explicit Key)

commit a0dcf3e982930ac69a6dc417b2df81e12587d255
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 14:33:26 2020 -0600

    Added Diff Tool

commit 7891cd9c4c5c28b77ccd0d04dca38c411d257489
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 14:04:41 2020 -0600

    Remote List/Snapshot Commands

commit 902ea36555d6bb2a296d5f371a43681850c05cc7
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 12:43:01 2020 -0600

    Registered Snapshot Command

commit f78f9f4296bbda52dd8e5eb51f93b24f249f05a9
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 12:42:47 2020 -0600

    Added Snapshot Command

commit 5aa187eedc8f353d58ee872dc555899bf30a41f8
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Sun Jul 5 11:41:01 2020 -0600

    Output Updates

commit 7c151abe76403f2f2d045fce510f6a6044bd0cd1
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Wed Jul 1 13:34:18 2020 -0600

    Prod Updates

commit 2f82b5f5039fa89bd615a6a0aa2f4b6835be592f
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Wed Jul 1 12:13:19 2020 -0600

    Bucket Validation and Reset Updates

commit 671a98b4e325757ebc700c7927fcf9ec07dbefaf
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Wed Jul 1 11:30:44 2020 -0600

    Prod updates

commit 3027495a9dbf4b89fc7ea93aa1139349f0bfbde1
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jun 30 11:11:55 2020 -0600

    Updated Settings.toString

commit 7c5bec745641ac29e46f875a0843017ce28ba377
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jun 30 11:07:12 2020 -0600

    Converted ExcludedPaths -> IncludedPaths

commit e103041c1a3a57335eb2505089fee885197f84c4
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Tue Jun 30 08:29:13 2020 -0600

    Migration Updates

commit 1fd0c96fabeaa2c614bc0a2ee148078b97ba579f
Author: Matthew Paiz <mpaiz@indicalab.com>
Date:   Mon Jun 29 12:17:21 2020 -0600

    Added Date Cuttoff

+++ BucketManager.cs
@@ -29,19 +29,17 @@
                 region: settings.RegionEndpoint);
 
             this.BucketName = settings.BucketName;
-            this.DriveLetter = settings.DriveLetter;
             this.PollingInterval = settings.PollingInterval;
             this.DebugMode = settings.DebugMode;
             this.MaxDownloads = settings.MaxDownloads;
-            this.ExcludedPaths = settings.ExcludedPaths ?? new List<string>();
+            this.IncludedPaths = settings.IncludedPaths ?? new List<string>();
+            this.Settings = settings;
             this.Logger = logger;
         }
 
         public TimeSpan PollingInterval { get; }
 
         public string BucketName { get; }
-
-        public char DriveLetter { get; }
 
         private AmazonS3Client Client { get; }
 
@@ -51,7 +49,9 @@
 
         private ILogger Logger { get; }
 
-        private IEnumerable<string> ExcludedPaths { get; }
+        private IEnumerable<string> IncludedPaths { get; }
+
+        private Settings Settings { get; }
 
         public async Task<IEnumerable<SourceImage>> ListPendingAsync(ServiceProvider provider)
         {
@@ -168,7 +168,9 @@
         public async Task ResetAsync(ServiceProvider provider)
         {
             var dbContext = provider.GetService<BucketMonitorContext>();
-            await dbContext.Database.ExecuteSqlRawAsync("TRUNCATE TABLE IMAGE");
+            var toDelete = dbContext.Image.Where(x => x.Bucket.Name == this.BucketName);
+            dbContext.Image.RemoveRange(toDelete);
+            await dbContext.SaveChangesAsync();
         }
 
         public async Task Summarize(ServiceProvider provider)
@@ -214,6 +216,20 @@
                 .AsNoTracking()
                 .ToListAsync())
                 .Select(x => this.GetFromCacheEntry(x));
+        }
+        
+        public async Task ValidateBucket(BucketMonitorContext context)
+        {
+           var bucket = await context.Bucket
+                .AsNoTracking()
+                .Where(x => x.Name == this.BucketName)
+                .SingleOrDefaultAsync();
+
+            if (bucket == null)
+            {
+                this.Logger.LogError("Bucket is not configured: {0}", this.BucketName);
+                throw new Exception();
+            }
         }
 
         private async Task<IDictionary<string, ImageEntry>> GetImageMapAsync(BucketMonitorContext context)
@@ -349,6 +365,7 @@
             int cacheAdditions = 0;
             int count = 0;
             ListObjectsV2Response response;
+            long total = 0;
             do
             {
                 using (var scoped = provider.CreateScope())
@@ -363,8 +380,11 @@
                     var images = new List<SourceImage>();
                     var toAdd = new List<ImageEntry>();
 
+                    total += response.S3Objects.Count();
+
                     var filtered = response.S3Objects
-                        .Where(x => x.Key.Length <= MAX_PATH_LENGTH);
+                        .Where(x => x.Key.Length <= MAX_PATH_LENGTH)
+                        .Where(x => this.IsIncludedPath(x.Key));
 
                     foreach (var obj in filtered)
                     {
@@ -390,11 +410,11 @@
 
                     if (cacheAdditions > 0)
                     {
-                        console.Update($"Loading Objects: {results.Count()}, CacheAdditions={cacheAdditions}");
+                        console.Update($"Loading Objects: {total}, CacheAdditions={cacheAdditions}");
                     }
                     else
                     {
-                        console.Update($"Loading Objects: {results.Count()}");
+                        console.Update($"Loading Objects: {total}");
                     }
 
                 }
@@ -411,14 +431,18 @@
         {
             return TryConvertPath(
                 key,
-                this.DriveLetter,
-                this.ExcludedPaths,
+                this.Settings.RootPath,
+                this.IncludedPaths,
                 out file);
         }
 
-        public static bool TryConvertPath(string key, char driveLetter, IEnumerable<string> excluded, out FileInfo file)
+        public bool IsIncludedPath(string key) => IsIncludedPath(key, this.IncludedPaths);
+
+        public static bool IsIncludedPath(string key, IEnumerable<string> included) => included.Count() == 0 || included.Any(x => key.StartsWith($"{x}/"));
+
+        public static bool TryConvertPath(string key, string rootPath, IEnumerable<string> included, out FileInfo file)
         {
-            if (excluded.Any(x => key.StartsWith($"{x}/")))
+            if (!IsIncludedPath(key, included))
             {
                 file = default(FileInfo);
                 return false;
@@ -430,7 +454,7 @@
             }
             else
             {
-                var path = Path.Combine($"{driveLetter}:\\", key);
+                var path = Path.Combine(rootPath, key);
                 file = new FileInfo(path); 
                 return true;
             }
@@ -448,46 +472,56 @@
 
             try
             {
-                var response = await this.Client.GetObjectAsync(new GetObjectRequest()
+                if (!file.Exists)
                 {
-                    BucketName = this.BucketName,
-                    Key = image.Key
-                });
-
-                var tmp = Path.GetTempFileName();
-
-                long transferred = 0;
-                response.WriteObjectProgressEvent += (sender, p) =>
-                {
-                    if (!started)
+                    var response = await this.Client.GetObjectAsync(new GetObjectRequest()
                     {
-                        tracker.Start();
+                        BucketName = this.BucketName,
+                        Key = image.Key
+                    });
+
+                    var tmp = Path.GetTempFileName();
+
+                    long transferred = 0;
+                    response.WriteObjectProgressEvent += (sender, p) =>
+                    {
+                        if (!started)
+                        {
+                            tracker.Start();
+                        }
+                        started = true;
+
+                        downloaded = p.TransferredBytes;
+                        tracker.Downloaded(downloaded - transferred);
+                        transferred = downloaded;
+                    };
+
+                    await response.WriteResponseStreamToFileAsync(tmp, false, CancellationToken.None);
+
+                    if (this.DebugMode)
+                    {
+                        File.Delete(tmp);
                     }
-                    started = true;
+                    else
+                    {
+                        if (!this.SafeMove(tmp, file.FullName))
+                        {
+                            image.MarkFailed();
+                            dbTracker.Update(entry, ImageStatus.Failed);
+                            tracker.Fail();
+                            return null;
+                        }
+                    }
 
-                    downloaded = p.TransferredBytes;
-                    tracker.Downloaded(downloaded - transferred);
-                    transferred = downloaded;
-                };
-
-                await response.WriteResponseStreamToFileAsync(tmp, false, CancellationToken.None);
-
-                if (this.DebugMode)
-                {
-                    File.Delete(tmp);
+                    this.Logger.LogDebug("Download Complete: {0} -> {1}", image.Key, file?.FullName ?? "-");
                 }
                 else
                 {
-                    if (!this.SafeMove(tmp, file.FullName))
-                    {
-                        image.MarkFailed();
-                        dbTracker.Update(entry, ImageStatus.Failed);
-                        tracker.Fail();
-                        return null;
-                    }
+                    this.Logger.LogDebug("Skipping Existing File: {0} -> {1}", image.Key, file?.FullName ?? "-");
+                    started = true;
+                    tracker.Start();
+                    tracker.Downloaded(image.TotalBytes);
                 }
-
-                this.Logger.LogDebug("Download Complete: {0} -> {1}", image.Key, file?.FullName ?? "-");
 
                 image.MarkCompleted();
                 dbTracker.Update(entry, ImageStatus.Completed);
@@ -524,23 +558,14 @@
                 var sourceFile = new FileInfo(source); 
                 var destinationFile = new FileInfo(destination); 
 
-                if (sourceFile.Length == destinationFile.Length)
+                if (!destinationFile.Directory.Exists)
                 {
-                    File.Move(source, destination, overwrite: true);
-                    return true;
-                }
-                else
-                {
-                    this.Logger.LogDebug(
-                        "Cancelling Move Operation. Source file length ({0}) does not match destination ({1}): {2}",
-                        sourceFile.Length,
-                        destinationFile.Length,
-                        destination);
-
-                    File.Delete(source);
-                    return false;
+                    Directory.CreateDirectory(destinationFile.Directory.FullName);
+                    this.Logger.LogDebug($"Created Directory: {destinationFile.Directory.FullName}");
                 }
 
+                File.Move(source, destination);
+                return true;
             }
             catch (Exception ec)
             {
